% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomVector_meanEstimation.R
\name{CG.mean}
\alias{CG.mean}
\title{The Catoni–Giulini mean estimator using R}
\usage{
CG.mean(matrix, alpha = 0.1)
}
\arguments{
\item{matrix}{Design matrix over \eqn{R^{p \times n}}, \code{n} is the number of random vectors, \code{p} is the dimension of random vectors}

\item{alpha}{A small and positive Catoni–Giulini parameter}
}
\value{
The Catoni–Giulini mean estimator of \code{n} column vectors of size \code{p} of the matrix
}
\description{
The Catoni–Giulini mean estimator using R
}
\examples{
\dontrun{
    random_matrix_normal <- matrix(rnorm(20000, mean = 0, sd = 1), nrow = 200, ncol = 100)
    CG.mean(random_matrix_normal)
}
}
