% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomVector_meanEstimation.R
\name{empirical.mean}
\alias{empirical.mean}
\title{The empirical mean using R}
\usage{
empirical.mean(matrix)
}
\arguments{
\item{matrix}{Design matrix over \eqn{R^{p \times n}}, \code{n} is the number of random vectors, \code{p} is the dimension of random vectors}
}
\value{
The empirical mean of \code{n} column vectors of size \code{p} of the matrix
}
\description{
The empirical mean using R
}
\examples{
\dontrun{
    random_matrix_normal <- matrix(rnorm(20000, mean = 0, sd = 1), nrow = 200, ncol = 100)
    empirical.mean(random_matrix_normal)
}
}
