% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomVector_meanEstimation.R
\name{gibbsSamplingR}
\alias{gibbsSamplingR}
\title{A Gibbs sampler using R}
\usage{
gibbsSamplingR(N, thin)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
rnR <- gibbsSamplingR(100,10)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
