% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomVector_meanEstimation.R
\name{MOM.mean}
\alias{MOM.mean}
\title{The median-of-mean estimator using R}
\usage{
MOM.mean(matrix, k, method = "geometric")
}
\arguments{
\item{matrix}{Design matrix over \eqn{R^{p \times n}}, \code{n} is the number of random vectors, \code{p} is the dimension of random vectors}

\item{k}{the number of groups you want after the data is evenly grouped}

\item{method}{the special definition of "median" you choose, such as coordinate-wise, Euclidean-ball and geometric}
}
\value{
The median-of-mean estimator of \code{n} column vectors of size \code{p} of the matrix
}
\description{
The median-of-mean estimator using R
}
\examples{
\dontrun{
    random_matrix_normal <- matrix(rnorm(20000, mean = 0, sd = 1), nrow = 200, ncol = 100)
    robust.mean(random_matrix_normal)
}
}
