% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{projected_gradient_descent}
\alias{projected_gradient_descent}
\title{High-Dimensional Robust Mean Estimation via Gradient Descent using Rcpp}
\usage{
projected_gradient_descent(matrix, epsilon, max_iter = 1000L, tol = 1e-05)
}
\arguments{
\item{matrix}{Design matrix over \eqn{R^{p \times n}}, \code{n} is the number of random vectors, \code{p} is the dimension of random vectors}

\item{epsilon}{Estimated proportion of outliers in a dataset}

\item{max_iter}{The maximum number of iterations of the algorithm}

\item{tol}{The convergence coefficient of the algorithm}
}
\value{
A weight vector for a weighted mean estimate
}
\description{
High-Dimensional Robust Mean Estimation via Gradient Descent using Rcpp
}
\examples{
\dontrun{
random_matrix_normal <- matrix(rnorm(20000, mean = 0, sd = 1), nrow = 200, ncol = 100)
w <- projected_gradient_descent(random_matrix_normal, epsilon)
mu_w <- random_matrix_normal \%*\% w
print("The weighted mean estimate:")
}
}
