% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomVector_meanEstimation.R
\name{robust.mean}
\alias{robust.mean}
\title{The mean estimator with robust loss functions using R}
\usage{
robust.mean(matrix, robust.par = 1, method = "huber")
}
\arguments{
\item{matrix}{Design matrix over \eqn{R^{p \times n}}, \code{n} is the number of random vectors, \code{p} is the dimension of random vectors}

\item{robust.par}{the parameter of the selected loss function}

\item{method}{the special robust loss function name you choose, such as huber, bisquare and wang}
}
\value{
The mean estimator with robust loss functions of \code{n} column vectors of size \code{p} of the matrix
}
\description{
The mean estimator with robust loss functions using R
}
\examples{
\dontrun{
    random_matrix_normal <- matrix(rnorm(20000, mean = 0, sd = 1), nrow = 200, ncol = 100)
    robust.mean(random_matrix_normal)
}
}
