\name{SA24204187-package}
\alias{SA24204187-package}
\docType{package}
\title{
Functions of kinds of mean estimators
}
\description{
Several R and Rcpp function are included in this package for mean estimation.
}
\author{
Jianjun Zhang

Maintainer: Jianjun Zhang <2309730578@qq.com>
}
\keyword{package}
