% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomVector_meanEstimation.R
\name{selectCG.mean}
\alias{selectCG.mean}
\title{The Catoni–Giulini mean estimator selected by MSE using R}
\usage{
selectCG.mean(matrix, delta_CE = 0.05, level = 10)
}
\arguments{
\item{matrix}{Design matrix over \eqn{R^{p \times n}}, \code{n} is the number of random vectors, \code{p} is the dimension of random vectors}

\item{delta_CE}{A small and positive parameter meaning a confident level like 0.05}

\item{level}{The level of precision of the selection}
}
\value{
The Catoni–Giulini mean estimator of \code{n} column vectors of size \code{p} of the matrix selected by MSE
}
\description{
The Catoni–Giulini mean estimator selected by MSE using R
}
\examples{
\dontrun{
    random_matrix_normal <- matrix(rnorm(20000, mean = 0, sd = 1), nrow = 200, ncol = 100)
    selectCG.mean(random_matrix_normal)
}
}
