% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomVector_meanEstimation.R
\name{trimmed.mean}
\alias{trimmed.mean}
\title{The simple trimmed mean estimator using R}
\usage{
trimmed.mean(matrix, epsilon = 0.1)
}
\arguments{
\item{matrix}{Design matrix over \eqn{R^{p \times n}}, \code{n} is the number of random vectors, \code{p} is the dimension of random vectors}

\item{epsilon}{Estimated proportion of outliers in a dataset}
}
\value{
The simple trimmed mean estimator of \code{n} column vectors of size \code{p} of the matrix
}
\description{
The simple trimmed mean estimator using R
}
\examples{
\dontrun{
    random_matrix_normal <- matrix(rnorm(20000, mean = 0, sd = 1), nrow = 200, ncol = 100)
    trimmed.mean(random_matrix_normal)
}
}
